refactor: use VersionedTransaction in benchmarks
