+++ bench/src/helpers.rs
@@ -1,14 +1,10 @@
 use anyhow::Context;
-use itertools::Itertools;
 use lazy_static::lazy_static;
-use rand::{distributions::Alphanumeric, prelude::Distribution, SeedableRng};
 use solana_rpc_client::nonblocking::rpc_client::RpcClient;
-use solana_sdk::compute_budget;
-use solana_sdk::instruction::AccountMeta;
+
 use solana_sdk::{
     commitment_config::CommitmentConfig,
     hash::Hash,
-    instruction::Instruction,
     message::Message,
     pubkey::Pubkey,
     signature::{Keypair, Signature},
@@ -17,10 +13,9 @@ use solana_sdk::{
     transaction::Transaction,
 };
 use std::path::PathBuf;
-use std::{str::FromStr, time::Duration};
+use std::time::Duration;
 use tokio::time::Instant;
 
-const MEMO_PROGRAM_ID: &str = "MemoSq4gqABAXKb96qnH8TysNcWxMyWCqXgDLGmfcHr";
 const WAIT_LIMIT_IN_SECONDS: u64 = 60;
 
 lazy_static! {
@@ -77,132 +72,4 @@ impl BenchHelper {
 
         Transaction::new(&[funded_payer], message, blockhash)
     }
-
-    pub fn generate_random_strings(
-        num_of_txs: usize,
-        random_seed: Option<u64>,
-        n_chars: usize,
-    ) -> Vec<Vec<u8>> {
-        let seed = random_seed.map_or(0, |x| x);
-        let mut rng = rand_chacha::ChaCha8Rng::seed_from_u64(seed);
-        (0..num_of_txs)
-            .map(|_| Alphanumeric.sample_iter(&mut rng).take(n_chars).collect())
-            .collect()
-    }
-
-    #[inline]
-    pub fn generate_txs(
-        num_of_txs: usize,
-        funded_payer: &Keypair,
-        blockhash: Hash,
-        random_seed: Option<u64>,
-        cu_price_micro_lamports: u64,
-    ) -> Vec<Transaction> {
-        let seed = random_seed.map_or(0, |x| x);
-        let mut rng = rand_chacha::ChaCha8Rng::seed_from_u64(seed);
-        (0..num_of_txs)
-            .map(|_| {
-                let random_bytes: Vec<u8> = Alphanumeric.sample_iter(&mut rng).take(10).collect();
-
-                Self::create_memo_tx_small(
-                    &random_bytes,
-                    funded_payer,
-                    blockhash,
-                    cu_price_micro_lamports,
-                )
-            })
-            .collect()
-    }
-
-    // note: there is another version of this
-    pub fn create_memo_tx_small(
-        msg: &[u8],
-        payer: &Keypair,
-        blockhash: Hash,
-        cu_price_micro_lamports: u64,
-    ) -> Transaction {
-        let memo = Pubkey::from_str(MEMO_PROGRAM_ID).unwrap();
-        let instruction = Instruction::new_with_bytes(memo, msg, vec![]);
-
-        let cu_request: Instruction =
-            compute_budget::ComputeBudgetInstruction::set_compute_unit_limit(14000);
-
-        let instructions = if cu_price_micro_lamports > 0 {
-            let cu_budget_ix: Instruction =
-                compute_budget::ComputeBudgetInstruction::set_compute_unit_price(
-                    cu_price_micro_lamports,
-                );
-            vec![cu_request, cu_budget_ix, instruction]
-        } else {
-            vec![cu_request, instruction]
-        };
-
-        let message = Message::new(&instructions, Some(&payer.pubkey()));
-        Transaction::new(&[payer], message, blockhash)
-    }
-
-    pub fn create_memo_tx_large(
-        msg: &[u8],
-        payer: &Keypair,
-        blockhash: Hash,
-        cu_price_micro_lamports: u64,
-    ) -> Transaction {
-        let accounts = (0..8).map(|_| Keypair::new()).collect_vec();
-        let memo = Pubkey::from_str(MEMO_PROGRAM_ID).unwrap();
-        let instruction = Instruction::new_with_bytes(
-            memo,
-            msg,
-            accounts
-                .iter()
-                .map(|keypair| AccountMeta::new_readonly(keypair.pubkey(), true))
-                .collect_vec(),
-        );
-
-        let instructions = if cu_price_micro_lamports > 0 {
-            let cu_budget_ix: Instruction =
-                compute_budget::ComputeBudgetInstruction::set_compute_unit_price(
-                    cu_price_micro_lamports,
-                );
-            vec![cu_budget_ix, instruction]
-        } else {
-            vec![instruction]
-        };
-
-        let message = Message::new(&instructions, Some(&payer.pubkey()));
-
-        let mut signers = vec![payer];
-        signers.extend(accounts.iter());
-
-        Transaction::new(&signers, message, blockhash)
-    }
-}
-
-#[test]
-fn transaction_size_small() {
-    let blockhash = Hash::default();
-    let payer_keypair = Keypair::from_base58_string(
-        "rKiJ7H5UUp3JR18kNyTF1XPuwPKHEM7gMLWHZPWP5djrW1vSjfwjhvJrevxF9MPmUmN9gJMLHZdLMgc9ao78eKr",
-    );
-
-    let seed = 42;
-    let random_strings = BenchHelper::generate_random_strings(1, Some(seed), 10);
-    let rand_string = random_strings.first().unwrap();
-    let tx = BenchHelper::create_memo_tx_small(rand_string, &payer_keypair, blockhash, 300);
-
-    assert_eq!(bincode::serialized_size(&tx).unwrap(), 231);
-}
-
-#[test]
-fn transaction_size_large() {
-    let blockhash = Hash::default();
-    let payer_keypair = Keypair::from_base58_string(
-        "rKiJ7H5UUp3JR18kNyTF1XPuwPKHEM7gMLWHZPWP5djrW1vSjfwjhvJrevxF9MPmUmN9gJMLHZdLMgc9ao78eKr",
-    );
-
-    let seed = 42;
-    let random_strings = BenchHelper::generate_random_strings(1, Some(seed), 232);
-    let rand_string = random_strings.first().unwrap();
-    let tx = BenchHelper::create_memo_tx_large(rand_string, &payer_keypair, blockhash, 300);
-
-    assert_eq!(bincode::serialized_size(&tx).unwrap(), 1222);
 }
